refactor: consolidate deploy-workflow verification onto one model
Fold the #921 workflow-gating suite and the merged #935
deploy-workflow-security suite onto a single workflow-as-data model in
adr_guard.py (the _dw_* helpers: a YAML loader and a constrained
if-expression evaluator), removing the two-homes / two-parsers split that
otherwise seeds fragmentation across the verification-stack epic.

- Promote runner exposure to a hard adr_guard check
  (deploy-workflow-runner-exposure) wired to the existing ADR-003-R5, which
  mandated the invariant but had no check. It evaluates each reusable
  workflow's self-hosted job if-expression for a pull_request event and
  requires it to fail closed - stronger than the prior substring check,
  which a guard broadened with `|| always()` would slip past.
- Move the remaining invariants (#781 upstream gating, #892 routing matrix,
  #913 filter split) plus #935's github_environment binding and ECR digest
  checks into one consolidated suite at
  scripts/adr_guard/tests/test_deploy_workflow.py, run by the existing
  Tests (adr_guard) job.
- Remove the standalone scripts/workflow_gating/ package, its Quality job,
  and test_deploy_workflow_security.py.
