+++ changelog.d/921.added.md
@@ -1,12 +1,17 @@
-**CI verifies the deploy control plane's gating invariants.** A new
-`workflow-gating-tests` Quality job parses `deploy.yml` and the reusable deploy
-workflows as data and asserts the gating rules that `actionlint` (syntax only)
-and `adr_guard` (narrow ADR checks) cannot infer: every deploy job fails closed
-when an upstream is `failure`/`cancelled` (#781), branch/event routing keeps
-`workflow_dispatch` on `main` as the only production-apply path and never
-deploys from `dev`/`main` (#892), the `portal_image` (app image) and
-`shifter_platform` (Terraform) change filters stay split (#913), and no
-`pull_request` event can reach a self-hosted apply/deploy job (the
-runner-exposure fix). A constrained GitHub Actions `if:` expression evaluator
-proves each denied case semantically rather than by substring match. The job
-runs on every PR with `contents: read` only — no secrets, no cloud access.
+**Deploy control-plane gating is verified by one workflow-as-data model, with
+a new ADR-003-R5 hard check.** `adr_guard.py` now carries a single model that
+reads `deploy.yml` and the reusable deploy workflows as data and evaluates
+their `if:` gates, branch/event routing, and change filters semantically. A new
+`deploy-workflow-runner-exposure` adr_guard check enforces ADR-003-R5 at commit
+time: every self-hosted deploy job must fail closed on `pull_request`, proven by
+evaluating the job's `if:` for a pull_request event rather than substring
+matching (so a guard broadened with `|| always()` is caught). The consolidated
+test suite (`scripts/adr_guard/tests/test_deploy_workflow.py`) exercises the same
+model for the remaining invariants: deploy jobs fail closed when an upstream is
+`failure`/`cancelled` (#781), `workflow_dispatch` on `main` is the only
+production-apply path and no `pull_request` routes a provider deploy (#892), the
+`portal_image` (app image) and `shifter_platform` (Terraform) change filters
+stay split (#913), mutating jobs bind a GitHub Environment, and the engine deploy
+pins an immutable ECR digest. This replaces the earlier substring-based
+`test_deploy_workflow_security.py`, folding deploy-workflow verification into one
+home and one parser.
