--- scripts/workflow_gating/__init__.py
@@ -1,1 +1,0 @@
-"""Workflow-gating verification for the deploy control plane (#921)."""
